+++ src/api.h
@@ -96,7 +96,7 @@ struct chart {
 typedef void (*analysis_func)(struct candle *cnds, size_t indx);
 
 void chart_create_object_line(struct candle *cnd, size_t start_idx,
-    size_t start_price, size_t end_idx, size_t end_price,
+    uint32_t start_price, size_t end_idx, uint32_t end_price,
     const char *function_name);
 
 void chart_create_object_text(
